% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlDbConnect-package.R
\docType{package}
\name{sqbDbConnect-package}
\alias{sqbDbConnect}
\alias{sqbDbConnect-package}
\title{sqbDbConnect: Reference class object to keep track of SQL database connections}
\description{
Reference class object to keep track of SQL database connections
}
\keyword{internal}
