% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLDbConnect.R
\docType{class}
\name{SQLDbConnect-class}
\alias{SQLDbConnect-class}
\alias{SQLDbConnect}
\title{SQLDbConnect class definition}
\description{
SQLDbConnect class definition
}
\section{Fields}{

\describe{
\item{\code{db_name}}{}
}}


\examples{
sql_con <- SQLDbConnect$new()
sql_con$connect("http://www.gpecdata.med.ubc.ca/OCV/ocv.sqlite")

}
