% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlDbConnect-package.R
\docType{package}
\name{sqlDbConnect-package}
\alias{sqlDbConnect}
\alias{sqlDbConnect-package}
\title{sqlDbConnect: Reference class object to keep track of SQL database connections}
\description{
Reference class object to keep track of SQL database connections
}
\keyword{internal}
